Update the demo for v0.14
Now all the examples should work. The biggest change was to the query example:
the whole query system was removed in favour of `DragRect()` so I updated it to
use that instead. Otherwise most changes come from fixing function names and
argument ordering. In some cases there were functions outside of a `BeginPlot()`
block that needed to be inside one, otherwise they would segfault.